// backend/src/shared/pipelines/enum.ts
export enum ProfitAndLossTypeIds {
    income = "10",
    otherIncome = "13",
    expense = "12",
    otherExpense = "14",
    costOfGoodsSold = "11",
}
export enum BalanceSheetTypeIds {
  Bank = "0",
  AccountsReceivable = "1",
  OtherCurrentAssets = "2",
  FixedAssets = "3",
  OtherAssets = "4",
  AccountsPayable = "5",
  CreditCard = "6",
  OtherCurrentLiabilities = "7",
  LongTermLiabilities = "8",
  Equity = "9",
}